# customer-billing/docs/swagger.yaml
basePath: /
definitions:
  handlers.ChargeBatchEntry:
    properties:
      batch_id:
        type: string
      billing_account_id:
        type: string
      billing_provider_report:
        $ref: '#/definitions/handlers.chargeBatchReportEntry'
      cloud_provider_report:
        $ref: '#/definitions/handlers.chargeBatchReportEntry'
      cloud_service_provider_id:
        type: string
      status:
        type: string
    type: object
...
  handlers.ResourceChargeEntry:
    properties:
      billing_currency:
        type: string
      cloud_service_provider_id:
        type: string
      line_item_count:
        type: integer
      resource_id:
        type: string
      resource_name:
        type: string
      resource_type:
        type: string
      service_category:
        type: string
      service_name:
        type: string
      total_billed_cost:
        type: number
    type: object
...
  handlers.chargeBatchReportEntry:
    properties:
      batch_signature:
        type: string
      billed_currency:
        type: string
      created_at:
        type: string
      merkle_root:
        type: string
      received_at:
        type: string
      total_cost:
        type: number
      total_items:
        type: integer
    type: object
...
host: localhost:8082
info:
  contact:
    name: Andrew Rutherfoord
  description: Mock Cloud Service Provider billing API implementing the FOCUS spec.
  title: CSP Mock API
  version: "1.0"
paths:
...
  /billing/accounts/{id}/charge-batches:
    get:
      description: Each charge batch is returned once, with the billing provider's
        report and the cloud service provider's own report of the same batch ID side
        by side (either may be missing), plus a status indicating whether they agree.
      parameters:
      - description: Billing account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.ChargeBatchEntry'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List charge batches for a billing account, merging the billing provider's
        and cloud provider's reports
      tags:
      - billing
...
  /billing/accounts/{id}/resource-charges:
    get:
      description: Aggregates billed cost per resource entirely from the FOCUS line
        items fetched directly from the CSP (not from anything reported by the billing
        provider), so each row also shows which cloud service provider it came from.
      parameters:
      - description: Billing account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.ResourceChargeEntry'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List aggregated billed cost per resource
      tags:
      - billing
...
securityDefinitions:
  BearerAuth:
    description: Bearer <account_id>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"